# openapi.yaml
openapi: 3.0.3
info:
  title: Grantly
  description: >
    A production-style OpenID Connect authorization server with a
    developer-facing client management API layered on top of the raw
    protocol. Social login (Google, GitHub), dynamic client registration
    (RFC 7591), and a session-based dashboard API for managing your own
    OAuth applications.
  version: 1.0.0
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: Local development
...
paths:
...
  /login:
    get:
      tags: [App]
      summary: Provider-picker login page (HTML)
      description: Renders Grantly's own login page listing enabled identity providers. Used both for third-party OIDC client logins (with an `interaction` param) and standalone visits.
      parameters:
        - name: interaction
          in: query
          required: false
          schema: { type: string }
      responses:
        '200':
          description: HTML login page
          content:
            text/html:
              schema: { type: string }
  /session/me:
    get:
      tags: [App]
      summary: Current session user
      description: >
        Plain session-cookie check, distinct from /me (which requires a
        Bearer access token). A frontend dashboard calls this on load to
        decide whether to show a login or authenticated view. CORS-scoped
        to FRONTEND_URL, credentials required.
      responses:
        '200':
          description: Logged-in user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '401':
          description: No active session
...
  /auth/external/{provider}:
    get:
      tags: [App]
      summary: Start upstream OAuth login
      description: >
        Redirects to the given identity provider's consent screen. Rate-limited
        to 10 requests/60s per IP. Must be opened as a browser navigation.
      parameters:
        - name: provider
          in: path
          required: true
          schema: { type: string, enum: [google, github] }
        - name: interaction_uid
          in: query
          required: false
          description: Only set when resuming a third-party OIDC interaction, never for Grantly's own dashboard login
          schema: { type: string }
      responses:
        '302':
          description: Redirect to the provider's consent screen
  /auth/external/{provider}/callback:
    get:
      tags: [App]
      summary: Upstream OAuth callback
      description: >
        Validates the `state` parameter (CSRF protection), exchanges the
        code with the provider, links or creates a user by verified email,
        establishes a session, then redirects either to the resumed OIDC
        interaction or to FRONTEND_URL/dashboard.
      parameters:
        - name: provider
          in: path
          required: true
          schema: { type: string }
        - name: code
          in: query
          required: true
          schema: { type: string }
        - name: state
          in: query
          required: true
          schema: { type: string }
      responses:
        '302':
          description: Redirect to /interaction/{uid} or the frontend dashboard
  /interaction/{uid}:
    get:
      tags: [App]
      summary: Resolve login/consent prompt
      description: Internal to the OIDC interaction flow - resolves the login and/or consent prompt for the given interaction.
      parameters:
        - name: uid
          in: path
          required: true
          schema: { type: string }
      responses:
        '302':
          description: Redirect back into the OIDC flow, or to /login if no session
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: client_id and client_secret as Basic Auth credentials (client_secret_basic)
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    ApiEnvelope:
      type: object
      properties:
        status: { type: string, enum: [success, error] }
        message: { type: string }
        data: {}
...